coredump: shell: enable printing of coredump
The coredump-shell print command doesn't work with the flash-partition
backend. When copying data from the coredump, the return-value is
checked for !=0, but the copy-command actually returns the number of
copied bytes on success.
Note that it works with the in-memory backend, because its
implementation of the copy command always return 0.

This fix changes the check of the return value to <0 for the shell
command, as well as the return-value of the in-memory's backend
copy-command to the number of copied bytes.

Signed-off-by: Christian Frey <christian.frey@embedded-insights.ch>